renamed to demo test case to exclude in build
